resolve circular dependencies - merge one.js and all.js
