+++ lib/handlers/blockquote.js
@@ -4,7 +4,7 @@
  */
 
 import {wrap} from '../wrap.js'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
