+++ lib/handlers/heading.js
@@ -3,7 +3,7 @@
  * @typedef {import('../index.js').Handler} Handler
  */
 
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
