+++ lib/handlers/link-reference.js
@@ -6,7 +6,7 @@
 
 import normalize from 'mdurl/encode.js'
 import {revert} from '../revert.js'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
