+++ lib/handlers/link.js
@@ -5,7 +5,7 @@
  */
 
 import normalize from 'mdurl/encode.js'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
