+++ lib/handlers/list-item.js
@@ -8,7 +8,7 @@
  */
 
 import {u} from 'unist-builder'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
