+++ lib/handlers/list.js
@@ -5,7 +5,7 @@
  */
 
 import {wrap} from '../wrap.js'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
