+++ lib/handlers/root.js
@@ -4,7 +4,7 @@
  */
 
 import {u} from 'unist-builder'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 import {wrap} from '../wrap.js'
 
 /**
