+++ lib/handlers/table.js
@@ -8,7 +8,7 @@
 
 import {pointStart, pointEnd} from 'unist-util-position'
 import {wrap} from '../wrap.js'
-import {all} from '../all.js'
+import {all} from '../traverse.js'
 
 /**
  * @type {Handler}
