+++ lib/index.js
@@ -67,7 +67,7 @@ import {visit} from 'unist-util-visit'
 import {pointStart, pointEnd} from 'unist-util-position'
 import {generated} from 'unist-util-generated'
 import {definitions} from 'mdast-util-definitions'
-import {one} from './one.js'
+import {one} from './traverse.js'
 import {footer} from './footer.js'
 import {handlers} from './handlers/index.js'
