+++ lib/revert.js
@@ -6,7 +6,7 @@
  */
 
 import {u} from 'unist-builder'
-import {all} from './all.js'
+import {all} from './traverse.js'
 
 /**
  * Return the content of a reference without definition as plain text.
