+++ lib/traverse.js
@@ -5,10 +5,10 @@
  * @typedef {import('mdast').Text} Text
  * @typedef {import('./index.js').H} H
  * @typedef {import('./index.js').Handler} Handler
+ * @typedef {import('./index.js').Content} Content
  */
 
 import {u} from 'unist-builder'
-import {all} from './all.js'
 
 var own = {}.hasOwnProperty
 
@@ -73,4 +73,46 @@ function text(node) {
 function returnNode(h, node) {
   // @ts-ignore pass through any unknown node.
   return node.children ? {...node, children: all(h, node)} : node
+}
+
+/**
+ * @param {H} h
+ * @param {Node} parent
+ */
+export function all(h, parent) {
+  /** @type {Array.<Node>} */
+  // @ts-ignore looks like a parent.
+  var nodes = parent.children || []
+  var index = -1
+  /** @type {Array.<Content>} */
+  var values = []
+  /** @type {Content} */
+  var result
+  /** @type {Content} */
+  var head
+
+  while (++index < nodes.length) {
+    // @ts-ignore looks like a parent.
+    result = one(h, nodes[index], parent)
+
+    if (result) {
+      if (index && nodes[index - 1].type === 'break') {
+        if (result.type === 'text') {
+          result.value = result.value.replace(/^\s+/, '')
+        }
+
+        if (result.type === 'element') {
+          head = result.children[0]
+
+          if (head && head.type === 'text') {
+            head.value = head.value.replace(/^\s+/, '')
+          }
+        }
+      }
+
+      values = values.concat(result)
+    }
+  }
+
+  return values
 }
