+++ package.json
@@ -78,7 +78,8 @@
     "rules": {
       "max-depth": "off",
       "no-var": "off",
-      "prefer-arrow-callback": "off"
+      "prefer-arrow-callback": "off",
+      "import/no-cycle": "error"
     }
   },
   "remarkConfig": {
